Flower/federated_aqi/requirements.txt

flower
hydra-core
ray
numpy
pandas
scikit-learn
tensorflow
